FIX(#56): 이미지가 빈 리스트일 때 공지사항, 알림장 이미지 삭제 방지

+++ src/main/kotlin/com/aimory/service/NoteService.kt
@@ -173,7 +173,11 @@ class NoteService(
             val imageUrls = note.noteImages.map {
                 it.imageUrl
             }
-            s3Service.deleteFiles(imageUrls)
+
+            // 이미지가 없을 경우 고려
+            if (imageUrls.isNotEmpty()) {
+                s3Service.deleteFiles(imageUrls)
+            }
 
             noteRepository.deleteById(note.id)
             deleteNoteIds.add(note.id)
