+++ src/main/kotlin/com/aimory/service/NoticeService.kt
@@ -117,7 +117,11 @@ class NoticeService(
             val imageUrls = notice.noticeImages.map {
                 it.imageUrl
             }
-            s3Service.deleteFiles(imageUrls)
+
+            // 이미지가 없을 경우 고려
+            if (imageUrls.isNotEmpty()) {
+                s3Service.deleteFiles(imageUrls)
+            }
 
             noticeRepository.deleteById(notice.id)
             deleteNoticeIds.add(notice.id)
